+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@ This document summarizes the most important changes of the individual releases o
 
 ## 2024.02
 
+* Add `synlig` (SystemVerilog plugin for Yosys)
 * Add Python packages for [IIC-RALF](https://github.com/iic-jku/IIC-RALF).
 * Add simple analog (inverter) and digital (counter) design examples in `/foss/examples`.
 * Add `libman` as a proposal for a design manager.